+++ src/main/java/com/ntselishchev/libraryapp/mapper/BookMapper.java
@@ -11,15 +11,15 @@ import java.sql.SQLException;
 public class BookMapper implements RowMapper<Book> {
     @Override
     public Book mapRow(ResultSet resultSet, int i) throws SQLException {
-        final int id = resultSet.getInt("id");
+        final long id = resultSet.getLong("id");
         final String title = resultSet.getString("title");
 
         Author author = new Author();
-        author.setId(resultSet.getInt("author_id"));
+        author.setId(resultSet.getLong("author_id"));
         author.setName(resultSet.getString("author_name"));
 
         Genre genre = new Genre();
-        genre.setId(resultSet.getInt("genre_id"));
+        genre.setId(resultSet.getLong("genre_id"));
         genre.setTitle(resultSet.getString("genre_title"));
 
         return new Book(id, title, author, genre);
